read snapshot table when index is unique, and add ordering_id to to_gbq
Using snapshot (time travel) ensures modifications to the underlying
table don't break our assumptions. It also would be surprising to the
user if the DataFrame was mutated by table updates.

Snapshot table not compatible with EXPORT DATA statement, so modified
I/O to first run a query (as it used to do). Since it was an easy part
of the refactor, add `ordering_id` option to `to_gbq` so folks can
serialize that if desired.

Change-Id: I6aa51ce83a21c5b9b0d01b4cdf31da099796bd9d