Check that types are specified in read_gbq_function
Functions created by the remote_function decorator have argument and
return types explicitly set, but UDFs in general may omit these. Omitted
return types are not compatible with Ibis, however, so we add a check
for this in read_gbq_function. We also test that read_gbq_function works
as expected for basic SQL and JavaScript UDFs and that types are
enforced.

We also fix a couple of miscellaneous issues: replacing asserts with
explicit exceptions, fixing a docstring format, and replacing
class/function imports with package/module imports.

Change-Id: Ic6fc1d51b582161970dcae5eadfba5c157e98411