feat: add series.sample (identical to existing dataframe.sample) (#187)
We're duplicating some arg-parsing logic here. Discussed briefly with Trevor. This is the case for other methods as well- we might want to add a sharing mechanism for dataframe/series (superclass like pandas?) in the future.

The documentation already exists in third_party/core/generic.py, which is actually what prompted this feat/fix.