chore: update toc entry to properly include summary pages (#415)
When the files aren't present, docfx does not convert the files from `.yml` extension to `.html`. I've tested this locally to ensure docfx keeps the file extension to `.html` as needed, which we'll need for the new files added for summary pages.

The only entry that should change in the future would be `summary_overview.html` to `summary_overview.md`. Filed #414  to keep track of that.

🦕