fix: remove pre-caching of remote function results
This will save redundant remote function execution as reported in
b/370088754. The trade-off is that any remote function integration
issues will be caughts only at a later point through a usage triggered
sql execution. Why the caching is not working as indended in the
reported use case in the bug? as per TrevorBergeron@ "the cached execution is useless when it is implicitly joined back to the base dataframe"